requirement.txt
streamlit>=1.24.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-pinecone>=0.0.1
langchain-community>=0.0.1
sentence-transformers>=2.2.0
pinecone-client>=2.2.0
google-generativeai>=0.3.0
Pillow>=9.0.0
GitPython>=3.1.0
aiofiles>=0.8.0
asyncio>=3.4.3
logging>=0.5.1.2